0.0

+++ elderlyManagement/src/com/xiaohe/service/impl/ProductServiceImpl.java
@@ -405,28 +405,30 @@ public List<ProductCustom> quertyNoBranchRecommendProduct(Integer branchid) {
 				e.printStackTrace();
 			}
 		}
-		if(productMapper.updateByPrimaryKeySelective(productInfo) < 0) return false;
+		if("".equals(productInfo.getStoragemethod())) productInfo.setStoragemethod(null);
+		if("".equals(productInfo.getEdiblemethod())) productInfo.setEdiblemethod(null);
+		if(productMapper.updateByPrimaryKeySelective(productInfo) <= 0) return false;
 		
 		
 		
 		Producttaste record = new Producttaste();
 		record.setProductid(productInfo.getProductid());
 		
-		if(productInfo.getTasteString() != null){
+		if(productInfo.getTasteString() != null && !"".equals(productInfo.getTasteString())){
 			if(producttasteMapper.deleteProducttasteByProductid(record.getProductid()) < 0) return false;
 			for(String temp : productInfo.getTasteString().split(" ")){
 				record.setProducttaste(temp);
-				if(producttasteMapper.insertSelective(record) < 0) return false;
+				if(producttasteMapper.insertSelective(record) <= 0) return false;
 			}
 		}
 		
 		Productcolour recordColor = new Productcolour();
 		recordColor.setProductid(productInfo.getProductid());
-		if(productInfo.getColorString() != null){
+		if(productInfo.getColorString() != null && !"".equals(productInfo.getColorString())){
 			if(productcolourMapper.deleteColorByProductId(recordColor.getProductid()) < 0) return false;
 			for(String temp : productInfo.getColorString().split(" ")){
 				recordColor.setProductcolour(temp);
-				if(productcolourMapper.insert(recordColor) < 0) return false;
+				if(productcolourMapper.insertSelective(recordColor) <= 0) return false;
 			}
 		}
